bug fix and code check
1. range: input float64-> output float32
2. convert pass for acl
3. ut fix

+++ tests/st/pi_jit/one_stage/test_base_op.py
@@ -1155,7 +1155,7 @@ def test_attr_as_inputs_7():
     cond = [Tensor([-1]), 1, 2.2, np.float16(3.3)]
     for i in cond:
         net.x = i
-        x = Tensor(np.random.rand(1))
+        x = Tensor(np.random.rand(1).astype(np.float32))
         y = net(x)
         assert (net.x + x) == y
